Split widget: batching panel
